Got visualization working
